+++ code/src/sixsq/nuvla/db/es/loader.clj
@@ -7,9 +7,10 @@
 
 (defn load
   "Creates an Elasticsearch client based on the Elasticsearch Java API. Takes
-   the configuration parameters from the environmental variables ES_HOST and
-   ES_PORT. These default to 'localhost' and '9200' if not specified."
+   the configuration parameters from the environmental variable ES_ENDPOINTS.
+   These default to 'localhost:9200' if not specified."
   []
-  (-> (esru/create-es-client)
-      esru/wait-for-cluster
-      esrb/->ElasticsearchRestBinding))
+  (let [client (-> (esru/create-es-client)
+                    esru/wait-for-cluster)
+        sniffer (esru/create-es-sniffer client)]
+    (esrb/->ElasticsearchRestBinding client sniffer)))