+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - Support bulk delete for more resources and fix operations for collection
   - Job - Stale job because of race condition fix
   - NuvlaBox - Simplify names of nuvlabox sub-resources at commission time
+  - Allow to conifgure ES client for ES cluster and use ES sniffer.
 
 ## [4.2.5] - 2019-10-10